+++ backend/src/main/java/com/sim_backend/websockets/client/package-info.java
@@ -1,0 +1,2 @@
+/** This package contains the backend ocpp. message logic for the EV Charger Simulator. */
+package com.sim_backend.websockets.client;
